Improve JavaDoc for Inventory, Fixes BUKKIT-2632.
This commit changes many of the JavaDoc return notes to to be more
descriptive and to better match the standard found in found in the
JavaDocs for the Java language. Additional changes were also made to the
body of some JavaDoc comments, primarily to standardize to the third
person preferred in Sun's guidelines.

This commit address BUKKIT-2632. A more detailed description can be
found at https://bukkit.atlassian.net/browse/BUKKIT-2632